LE_Kotlin: renamed & moved ForPrinting.kt into constants
