+++ Attt_LocalEngine_Kotlin/src/main/kotlin/logic/GameField.kt
@@ -2,6 +2,9 @@ package logic
 
 import constants.MAX_GAME_FIELD_SIDE_SIZE
 import constants.MIN_GAME_FIELD_SIDE_SIZE
+import constants.SYMBOL_FOR_ABSENT_MARK
+import constants.SYMBOL_FOR_DIVIDER
+import constants.SYMBOL_FOR_NEW_LINE
 import elements.*
 import publicApi.AtttPlayer
 import utilities.Log
