+++ Attt_LocalEngine_Kotlin/src/main/kotlin/logic/PlayerProvider.kt
@@ -3,9 +3,9 @@ package logic
 import constants.MAX_NUMBER_OF_PLAYERS
 import constants.MIN_NUMBER_OF_PLAYERS
 import elements.Player
-import elements.SYMBOL_FOR_PLAYER_NONE
-import elements.SYMBOL_FOR_PLAYER_O
-import elements.SYMBOL_FOR_PLAYER_X
+import constants.SYMBOL_FOR_PLAYER_NONE
+import constants.SYMBOL_FOR_PLAYER_O
+import constants.SYMBOL_FOR_PLAYER_X
 import publicApi.AtttPlayer
 import utilities.Log
